import modules parsing to support py-env dependencies
The new module extends the ModuleFinder from find-imports
(in graveyard) to detect all the dependencies of the
file being converted.

Currently the parser is able to collect either
external packages (limited to those supported by Pyodide)
and any local module to be included in `paths`.

Local submodules (node.level > 0) not yet supported.