Improved parsing with local namespace inspection
Code parsing has been completely refactored and improved
to support the identification of corner import conditions
that won't work in Pyscript.
This includes unsupported packages, and local modules/packages
import that won't work if not imported directly.

As a result, ModuleFinder gathers all the packages via NamespaceInfo
(populated via pkgutil) and collects results in a FinderResult
object.
This object will be later used to specialise the processing result.